Prepare files
