+++ config/splat.us.stdre.yaml
@@ -63,6 +63,7 @@ segments:
       - [0x11A64, c]
       - [0x11A64, c, succubus]
       - [0x14774, c]
+      - [0x161DC, c]
       - [0x1697C, c]
       - [0x16F90, c, random]
       - [0x16FC0, c, update]
